Modified how natural attack damage dice are calcualted.
Instead of trying to find a fitting damage die, the monster's original damage die will now be given preference, unless the target damage is so low that a single die would be too high or the target damage is so high that it would result in an excessive number of dice.

Also remembered to do a prod build this time.